Arm Comms: Added support for 16-bit integers for position array.

+++ embedModules/arm2016/new/arm2016_types.h
@@ -13,7 +13,7 @@ enum Ecommand_type {MANUAL, INVERSE_KIN};
 
 struct packet {
 	Ecommand_type type;
-	int8_t position[3]; //x, y, z
+	int16_t position[3]; //x, y, z
 	int8_t duty_cycle[6]; // speed and direction of each motor
 	uint16_t checksum();
 };
